Product_search/requirement.txt
aiohttp==3.7.4.post0
arabic-reshaper==2.1.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==3.3.1
cycler==0.11.0
distlib==0.3.2
Django==3.0.5
django-braces==1.13.0
django-ckeditor==6.1.0
django-common-helpers==0.9.2
django-cors-headers==3.6.0
django-cron==0.5.1
django-cryptographic-fields==0.5.1
django-dynamic-resource-admin==0.1.1
django-js-asset==1.2.2
django-live-support==0.1.9
django-mathfilters==1.0.0
django-user-agents==0.4.0
Django-Verify-Email==0.0.5
django-widget-tweaks==1.4.8
django-with-extra-context-admin==0.1.0
djangorestframework==3.12.2
filelock==3.0.12
fonttools==4.29.1
future==0.18.2
geoip2==4.1.0
html5lib==1.1
idna==2.10
idna-ssl==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==8.1.0
platformdirs==2.1.0
pyactiveresource==2.2.1
pycparser==2.20
pycryptodome==3.9.9
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2020.5
PyYAML==5.3.1
reportlab==3.6.6
requests==2.25.1
ShopifyAPI==8.1.0
simplejson==3.17.2
six==1.15.0
sorl-thumbnail==12.8.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
virtualenv==20.6.0
webencodings==0.5.1
wget==3.2
xhtml2pdf==0.2.5
yarl==1.6.3
